changement type variable

+++ pointers_arrays_strings/100-atoi.c
@@ -9,7 +9,7 @@
 int _atoi(char *s)
 {
 	int i = 0;
-	int result = 0;
+	unsigned long result = 0;
 	int sign = 1;
 	int find = 0;
